Add OrderSettlementBuilder struct with some methods

+++ casa/src/domain/order_settlement.rs
@@ -1,9 +1,10 @@
 use chrono::prelude::*;
 use chrono::prelude::{DateTime, Utc};
 use domain::{SettlementStatus, Order};
+use db::{TryFromRow, TryFromRowError};
 
 pub struct OrderSettlement {
-    pub id: Option<u64>,
+    pub id: Option<i32>,
     pub started_at: DateTime<Utc>,
     pub settled_at: Option<DateTime<Utc>>,
     pub settlement_status: SettlementStatus,
